+++ app/build.gradle
@@ -46,7 +46,9 @@ android {
 }
 
 dependencies {
+    def nav_version = "2.5.3"
 
+    implementation("androidx.navigation:navigation-compose:$nav_version")
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation "androidx.compose.ui:ui:$compose_version"
     implementation "androidx.compose.material:material:$compose_version"
