+++ app/src/main/java/com/example/trackr_mobile/ui/Screen.kt
@@ -1,0 +1,6 @@
+package com.example.trackr_mobile.ui
+
+enum class Screen(val route: String) {
+    LOGIN("login"),
+    HOME("home"),
+}
